--- src/main/java/model/AppCommand.java
@@ -1,9 +1,0 @@
-package model;
-
-public enum AppCommand {
-    CREATE_TASK,
-    CLOSE_TASK,
-    DELETE_TASK,
-    REOPEN_TESK,
-    LIST_TASK
-}
