--- src/main/java/model/TaskStatus.java
@@ -1,7 +1,0 @@
-package model;
-
-public enum TaskStatus {
-    CREATED,
-    CANCELED,
-    DELETED
-}
